The only thing this is missing is distortion correction
I cleaned up as much as I could, but I couldn’t completely clean it
because I did not write the code in the first place. Distance should
work if camera constants are set correctly, but angle differential is
just being scaled based on targetX and imageWidth. We may want to
actually use the distance for that.